backend/requirements.txt
asgiref==3.8.1
channels==4.2.0
<<<<<<< HEAD
channels_redis==4.2.1
click==8.1.8
colorama==0.4.6
decorator==5.1.1
=======
>>>>>>> 097572a9b26d0de8d5f2cac76cb8430959a6088f
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
<<<<<<< HEAD
imageio==2.36.1
imageio-ffmpeg==0.5.1
joblib==1.4.2
moviepy==2.1.2
msgpack==1.1.0
nltk==3.9.1
numpy==2.2.1
pillow==10.4.0
proglog==0.1.10
psycopg2==2.9.10
pydub==0.25.1
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tqdm==4.67.1
=======
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
>>>>>>> 097572a9b26d0de8d5f2cac76cb8430959a6088f
tzdata==2024.2
